refactor: organizing ‘slots’ folder
This commit organizes the slots folder into a hierarchy, splitting up types, hooks, and utils into sub-folders based on the type of slot and operation

The basic organization is now:

slots
— ui
—— layout
—— widget
——— iframe

The entire structure properly exports from its children using index.ts files so that everything can be imported from ‘runtime’ without reaching into the package.

This commit also simplifies and organizes some of our hooks and helper functions to remove extraneous words.  We now have “useOperations”, “useWidgets”, “useLayout”, etc., rather than things like “useSlotWidgets”
