+++ runtime/index.ts
@@ -133,6 +133,4 @@ export {
   snakeCaseObject
 } from './utils';
 
-export type * from './slots';
-
 export * from './slots';
