+++ runtime/slots/index.ts
@@ -1,11 +1,4 @@
+export * from './hooks';
 export type * from './types';
-export {
-  LayoutOperationTypes,
-  WidgetOperationTypes
-} from './types';
-
-export { default as DefaultSlotLayout } from './DefaultSlotLayout';
-export { default as FederatedWidget } from './FederatedWidget';
-export * from './iframes';
-export { default as Slot } from './Slot';
-export { default as SlotContext } from './SlotContext';
+export * from './ui';
+export * from './utils';
