+++ runtime/slots/ui/hooks.ts
@@ -1,0 +1,6 @@
+import { useContext } from 'react';
+import SlotContext from './SlotContext';
+
+export function useSlotContext() {
+  return useContext(SlotContext);
+}
