+++ runtime/slots/ui/widget/index.ts
@@ -1,0 +1,7 @@
+export { default as FederatedWidget } from './FederatedWidget';
+export * from './hooks';
+export * from './iframe';
+export type * from './types';
+export { WidgetOperationTypes } from './types';
+export * from './utils';
+export { default as WidgetContext } from './WidgetContext';
