+++ shell/footer/CenterLinks.tsx
@@ -1,7 +1,7 @@
-import { useSlotWidgets } from '../../runtime/slots/data/hooks';
+import { useWidgets } from '../../runtime';
 
 export default function CenterLinks() {
-  const widgets = useSlotWidgets();
+  const widgets = useWidgets();
 
   return (
     <div className="d-flex flex-wrap column-gap-6 row-gap-4">
