+++ shell/footer/RightLinks.tsx
@@ -1,7 +1,7 @@
-import { useSlotWidgets } from '../../runtime/slots/data/hooks';
+import { useWidgets } from '../../runtime';
 
 export default function RightLinks() {
-  const widgets = useSlotWidgets();
+  const widgets = useWidgets();
 
   return (
     <div className="d-flex flex-column gap-3 align-items-end flex-grow-1 justify-content-between">
