+++ shell/header/mobile/MobileNavLinks.tsx
@@ -1,5 +1,5 @@
 import { Nav } from '@openedx/paragon';
-import Slot from '../../../runtime/slots/Slot';
+import { Slot } from '../../../runtime';
 
 export default function MobileNavLinks() {
   return (
