+++ shell/menus/NavDropdownMenuSlot.tsx
@@ -1,6 +1,7 @@
 import { NavDropdown } from '@openedx/paragon';
 import { useIntl } from 'react-intl';
-import { useLayoutOptionsById, useSlotWidgetsById } from '../../runtime/slots/data/hooks';
+
+import { useLayoutOptionsForId, useWidgetsForId } from '../../runtime';
 import {
   MenuItemName
 } from '../../types';
@@ -15,8 +16,8 @@ interface DropdownMenuSlotProps {
 
 export default function DropdownMenuSlot({ id, label }: DropdownMenuSlotProps) {
   const intl = useIntl();
-  const options = useLayoutOptionsById(id);
-  const widgets = useSlotWidgetsById(id);
+  const options = useLayoutOptionsForId(id);
+  const widgets = useWidgetsForId(id);
 
   const finalLabel = getItemLabel(options.label ?? label, intl);
